refactor: Decompose LabelItem component
Decompose LabelItem component into smaller, more focused components to enhance
readability, testability, and maintainability. Specific changes include:

- Abstract conditional className logic within LabelItem.
- Introduce LabelItemButtonContent as an isolated component.
- Move LabelItemDropdown component into LabelItem folder for better
organization.
- Introduce DropdownContentOnClose as a specialized component within
labelItemDropdown.
