+++ components/label/labelList/labelItem/labelItemDropdown/index.tsx
@@ -1,13 +1,13 @@
 import { ICON_DELETE, ICON_EDIT_NOTE } from '@data/materialSymbols';
 import { Types } from 'lib/types';
-import { Dropdown } from './dropdown';
-import { DropdownMenuItem } from './dropdown/dropdownMenuItem';
 import { optionsDropdownLabelItem } from '@options/misc';
 import { ActiveDropdownMenuItemEffect } from '@effects/activeDropdownMenuItemEffect';
 import { useLabelModalStateOpen } from '@hooks/modals';
 import { TypesOptionsDropdown } from '@lib/types/options';
 import { TypesLabel } from '@label/label.types';
 import { useLabelRemoveItem } from '@label/label.hooks';
+import { Dropdown } from '@dropdowns/v1/dropdown';
+import { DropdownMenuItem } from '@dropdowns/v1/dropdown/dropdownMenuItem';
 
 type Props = { options: TypesOptionsDropdown } & Pick<TypesLabel, 'label'> &
   Partial<Pick<Types, 'menuContentOnClose'>>;
